pyproject.toml
[project]
name = "video-agent"
version = "0.1.0"
description = "ArduinoVision - AI-powered hardware debugging agent"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "python-dotenv>=1.2.1",
    "vision-agents[deepgram,elevenlabs,gemini,getstream,huggingface,nvidia,openai,ultralytics]>=0.3.7",
    "pyserial>=3.5",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "websockets>=13.0",
    "pyjwt>=2.8.0",
]
